[Refactor] PostReview 리팩토링
