+++ src/pages/searchHome/PostReviewCenter.jsx
@@ -1,14 +1,15 @@
 import React from 'react'
-import { PostReview } from './PostReview'
+import { PostReview } from '../../component/PostReview'
 import { useLocation } from 'react-router-dom'
 
-export const ReviewCenter = () => {
+// 센터 리뷰
+export const PostReviewCenter = () => {
     const location = useLocation();
     const {center_name} = location.state || "";
 
   return (
     <>
-        <PostReview title={center_name} isRequired={0}></PostReview>
+        <PostReview title={center_name} type={'center'}></PostReview>
     </>
   )
 }
